requirements.txt
Flask==1.1.2
requests==2.26.0
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
Werkzeug==2.0.1
SQLAlchemy==1.4.20
APScheduler==3.7.0
python_dateutil==2.8.2
PyYAML==5.4.1
